pyproject.toml
[project]
name = "api-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "google-genai>=1.2.0",
    "openai>=1.63.0",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
    "twilio>=9.4.5",
    "uvicorn>=0.34.0",
]



[project.scripts]
start = "main:start"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["*.py", "/tests"]
exclude = [
  "*.json",
  "**/_compat.py",
]

[dependency-groups]
dev = [
    "pyright>=1.1.394",
    "pytest>=8.3.4",
    "ruff>=0.6.7",
]


[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212", "FA", "COM812", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402", "S101"]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
verboseOutput = true
